change the simple example module to a package
