+++ docs/example/accessors.py
@@ -1,32 +1,4 @@
-class CachedAccessor:
-    def __init__(self, name, accessor):
-        self._name = name
-        self._accessor = accessor
-
-    def __get__(self, obj, cls):
-        if obj is None:
-            return self._accessor
-
-        accessor_obj = self._accessor(obj)
-        setattr(obj, self._name, accessor_obj)
-
-        return accessor_obj
-
-
-class Example:
-    """test class"""
-
-    def __init__(self, data):
-        self._data = data
-
-
-def register_accessor(name):
-    def func(accessor):
-        setattr(Example, name, CachedAccessor(name, accessor))
-
-        return accessor
-
-    return func
+from example.extensions import CachedAccessor, register_accessor
 
 
 @register_accessor("test")
