+++ docs/example/object.py
@@ -1,0 +1,5 @@
+class Example:
+    """test class"""
+
+    def __init__(self, data):
+        self._data = data
